Remove unused views
These views and are not actually used by the `NinjaAPI`.
